+++ NEWS.md
@@ -20,6 +20,8 @@
   are co-linear (version `0.3.0.9004`)
 * Add support for `order > 2` for `fm_fem.fm_mesh_1d()` (version `0.3.0.9011`)
   and generally for `fm_matern_precision()` (version `0.3.0.9012`)
+* Add argument `units` to `fm_crs()` to allow setting the length unit
+  for the CRS on creation/extraction (version `0.3.0.9013`)
 
 ## Bug fixes